Provide defaults for LDAP mappings
git-svn-id: https://svn.calendarserver.org/repository/calendarserver/CalendarServer/trunk@15165 e27351fd-9f3e-4f54-a53b-843176b1656c

+++ contrib/performance/loadtest/sim.py
@@ -285,7 +285,8 @@ class LoadSimulator(object):
             if 'clientDataSerialization' in config:
                 serializationPath = config['clientDataSerialization']['Path']
                 if not config['clientDataSerialization']['UseOldData']:
-                    shutil.rmtree(serializationPath)
+                    if isdir(serializationPath):
+                        shutil.rmtree(serializationPath)
                 serializationPath = config['clientDataSerialization']['Path']
                 if not isdir(serializationPath):
                     try:
@@ -412,6 +413,7 @@ class LoadSimulator(object):
         return CalendarClientSimulator(
             self.records,
             populator,
+            Random(),
             self.parameters,
             self.reactor,
             self.server,
